fix(SadTalker): make sure make_annimation works for torch dtype of float16 as well

+++ src/sadtalker/src/facerender/modules/keypoint_detector.py
@@ -89,7 +89,9 @@ class KPDetector(nn.Module):
         shape = heatmap.shape
         heatmap = heatmap.unsqueeze(-1)
         grid = (
-            make_coordinate_grid(shape[2:], heatmap.device).unsqueeze_(0).unsqueeze_(0)
+            make_coordinate_grid(shape[2:], heatmap.device, dtype=heatmap.dtype)
+            .unsqueeze_(0)
+            .unsqueeze_(0)
         )
         value = (heatmap * grid).sum(dim=(2, 3, 4))
         kp = {"value": value}
